+++ CSETWebNg/src/app/app.module.ts
@@ -76,7 +76,13 @@ import { MatIconModule } from '@angular/material/icon';
 import { MatInputModule } from '@angular/material/input';
 import { MatListModule } from '@angular/material/list';
 import { MatMenuModule } from '@angular/material/menu';
-import { DateAdapter, MAT_DATE_FORMATS, MAT_DATE_LOCALE, MatNativeDateModule, MatRippleModule } from '@angular/material/core';
+import {
+  DateAdapter,
+  MAT_DATE_FORMATS,
+  MAT_DATE_LOCALE,
+  MatNativeDateModule,
+  MatRippleModule
+} from '@angular/material/core';
 import { MatPaginatorModule } from '@angular/material/paginator';
 import { MatProgressBarModule } from '@angular/material/progress-bar';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
@@ -384,7 +390,6 @@ import { CrrSummaryResultsComponent } from './assessment/results/crr/crr-summary
 import { CrrResultsPage } from './assessment/results/crr/crr-results-page/crr-results-page.component';
 import { CrrResultsDetailComponent } from './assessment/results/crr/crr-results-detail/crr-results-detail.component';
 import { CrrHeatmapComponent } from './assessment/results/crr/crr-heatmap/crr-heatmap.component';
-import { CrrService } from './services/crr.service';
 import { CmuService } from './services/cmu.service';
 import { Utilities } from './services/utilities.service';
 import { NCUAService } from './services/ncua.service';
@@ -612,667 +617,670 @@ import { MalcolmUploadErrorComponent } from './dialogs/malcolm/malcolm-upload-er
 import { FooterService } from './services/footer.service';
 import { AssessmentConvertCfComponent } from './assessment/prepare/assessment-info/assessment-convert-cf/assessment-convert-cf.component';
 import {
-    MAT_MOMENT_DATE_FORMATS,
-    MomentDateAdapter,
-    MAT_MOMENT_DATE_ADAPTER_OPTIONS,
+  MAT_MOMENT_DATE_FORMATS,
+  MomentDateAdapter,
+  MAT_MOMENT_DATE_ADAPTER_OPTIONS
 } from '@angular/material-moment-adapter';
 import { IseWarningsComponent } from './assessment/results/reports/ise-warnings/ise-warnings.component';
 //   import {DateAdapter, MAT_DATE_FORMATS, MAT_DATE_LOCALE} from '@angular/material/core';
 import { TrendCompareCompatibilityComponent } from './aggregation/trend-analytics/trend-compare-compatibility/trend-compare-compatibility.component';
 
 @NgModule({
-    imports: [
-        BrowserModule,
-        BrowserAnimationsModule,
-        FormsModule,
-        HttpClientModule,
-        CommonModule,
-        AppRoutingModule,
-        // Material
-        A11yModule,
-        CdkAccordionModule,
-        ClipboardModule,
-        CdkStepperModule,
-        CdkTableModule,
-        CdkTreeModule,
-        DragDropModule,
-        MatAutocompleteModule,
-        MatBadgeModule,
-        MatBottomSheetModule,
-        MatButtonModule,
-        MatButtonToggleModule,
-        MatCardModule,
-        MatCheckboxModule,
-        MatChipsModule,
-        MatStepperModule,
-        MatDatepickerModule,
-        MatDialogModule,
-        MatDividerModule,
-        MatExpansionModule,
-        MatGridListModule,
-        MatIconModule,
-        MatInputModule,
-        MatListModule,
-        MatMenuModule,
-        MatNativeDateModule,
-        MatPaginatorModule,
-        MatProgressBarModule,
-        MatProgressSpinnerModule,
-        MatRadioModule,
-        MatRippleModule,
-        MatSelectModule,
-        MatSidenavModule,
-        MatSliderModule,
-        MatSlideToggleModule,
-        MatSnackBarModule,
-        MatSortModule,
-        MatTableModule,
-        MatTabsModule,
-        MatToolbarModule,
-        MatTooltipModule,
-        MatTreeModule,
-        OverlayModule,
-        PortalModule,
-        ScrollingModule,
-        // AutosizeModule,
-        // NgChartsModule,
-        // MatButtonModule,
-        // MatToolbarModule,
-        // MatChipsModule,
-        // MatSlideToggleModule,
-        // MatInputModule,
-        // MatCardModule,
-        // MatSliderModule,
-        // MatDatepickerModule,
-        // MatNativeDateModule,
-        // MatFormFieldModule,
-        // MatSortModule,
-        // MatExpansionModule,
-        // MatAutocompleteModule,
-        // MatDialogModule,
-        // MatTooltipModule,
-        // MatSnackBarModule,
-        // MatSidenavModule,
-        // MatTreeModule,
-        // MatIconModule,
-        // MatDividerModule,
-        // MatProgressSpinnerModule,
-        // MatProgressBarModule,
-        // MatListModule,
-        // MatMenuModule,
-        // MatTabsModule,
-        FlexLayoutModule,
-        ReactiveFormsModule,
-        NgxSliderModule,
-        // TextareaAutosizeModule,
-        FileUploadModule,
-        AngularEditorModule,
-        RouterModule,
-        CurrencyMaskModule,
-        TranslocoRootModule,
-        NgbModule,
-        NgxChartsModule,
-        TooltipModule,
-        SwiperModule,
-        EllipsisModule,
-        HotkeyModule.forRoot(),
-        CodeEditorModule.forRoot({
-            typingsWorkerUrl: 'assets/workers/typings-worker.js',
-            baseUrl: 'assets/monaco'
-        }),
-    ],
-    declarations: [
-        AppComponent,
-        InitialComponent,
-        LoginComponent,
-        MyAssessmentsComponent,
-        AssessmentComponent,
-        ContactItemComponent,
-        PrepareComponent,
-        AssessmentInfoComponent,
-        AssessmentDetailComponent,
-        AssessmentContactsComponent,
-        AssessmentDemographicsComponent,
-        AssessmentInfoNcuaComponent,
-        AssessmentDetailNcuaComponent,
-        ResultsComponent,
-        SalSimpleComponent,
-        StandardsComponent,
-        ResetPassComponent,
-        EmailComponent,
-        ConfirmEqualValidatorDirective,
-        EmailValidatorDirective,
-        FocusDirective,
-        ZipCodeDirective,
-        AutoSizeDirective,
-        DigitsOnlyDirective,
-        RunScriptsDirective,
-        SalGenComponent,
-        SalNistComponent,
-        SalsComponent,
-        RegisterComponent,
-        EjectionComponent,
-        ChangePasswordComponent,
-        EditUserComponent,
-        AlertComponent,
-        ConfirmComponent,
-        FrameworkComponent,
-        RequiredDocsComponent,
-        IRPComponent,
-        ExamProfileComponent,
-        DiagramComponent,
-        AcetDetailComponent,
-        AboutComponent,
-        AdvisoryComponent,
-        QuestionsComponent,
-        QuestionBlockComponent,
-        QuestionExtrasComponent,
-        ResourceLibraryComponent,
-        OkayComponent,
-        ObservationsComponent,
-        IssuesComponent,
-        MeritCheckComponent,
-        SafePipe,
-        LinebreakPipe,
-        CompletionCountPipe,
-        LinebreakPlaintextPipe,
-        NullishCoalescePipe,
-        StatusCreateComponent,
-        ProgressComponent,
-        InViewComponent,
-        AnalysisComponent,
-        ReportsComponent,
-        OverviewComponent,
-        StandardsSummaryComponent,
-        StandardsRankedComponent,
-        StandardsResultsComponent,
-        ComponentsSummaryComponent,
-        ComponentsRankedComponent,
-        ComponentsResultsComponent,
-        ComponentsTypesComponent,
-        ComponentsWarningsComponent,
-        DashboardComponent,
-        RankedQuestionsComponent,
-        FeedbackComponent,
-        EnableProtectedComponent,
-        TermsOfUseComponent,
-        QuestionFiltersComponent,
-        AssessmentDocumentsComponent,
-        InlineParameterComponent,
-        GlobalParametersComponent,
-        ImportComponent,
-        UploadExportComponent,
-        KeyboardShortcutsComponent,
-        LicenseComponent,
-        AcetDashboardComponent,
-        AdminComponent,
-        SetListComponent,
-        CustomSetComponent,
-        RequirementListComponent,
-        QuestionListComponent,
-        BuilderBreadcrumbsComponent,
-        AddQuestionComponent,
-        RequirementDetailComponent,
-        AddRequirementComponent,
-        StandardDocumentsComponent,
-        RefDocumentComponent,
-        DomainMaturityFilterComponent,
-        IrpSummaryComponent,
-        ExamProfileSummaryComponent,
-        DiagramInventoryComponent,
-        DiagramInfoComponent,
-        DiagramComponentsComponent,
-        LinksComponent,
-        NetworkWarningsComponent,
-        ShapesComponent,
-        TextComponent,
-        ZonesComponent,
-        ExcelExportComponent,
-        NcuaExcelExportComponent,
-        MergeComponent,
-        MergeQuestionDetailComponent,
-        AggregationHomeComponent,
-        AliasAssessmentsComponent,
-        AggregationDetailComponent,
-        TrendAnalyticsComponent,
-        CompareAnalyticsComponent,
-        SelectAssessmentsComponent,
-        CompareSummaryComponent,
-        CompareMissedComponent,
-        CompareIndividualComponent,
-        CompareBestworstComponent,
-        ComponentOverrideComponent,
-        ExcelExportComponent,
-        LayoutBlankComponent,
-        LayoutSwitcherComponent,
-        LayoutMainComponent,
-        AcetLayoutMainComponent,
-        ReportTestComponent,
-        SiteDetailComponent,
-        ObservationTearoutsComponent,
-        EvalAgainstStandardsComponent,
-        ExecutiveSummaryComponent,
-        SecurityplanComponent,
-        SiteSummaryComponent,
-        TrendReportComponent,
-        TrendCompareCompatibilityComponent,
-        CompareReportComponent,
-        Assessment2InfoComponent,
-        ModelSelectComponent,
-        AssessmentConfigComponent,
-        CmmcAComponent,
-        CmmcLevelsComponent,
-        CmmcLevelResultsComponent,
-        CmmcLevelDrilldownComponent,
-        CmmcComplianceComponent,
-        CmmcGapsComponent,
-        Cmmc2LevelResultsComponent,
-        Cmmc2DomainResultsComponent,
-        SprsScoreComponent,
-        ComplianceScoreComponent,
-        ModelSelectComponent,
-        CategoryBlockComponent,
-        AskQuestionsComponent,
-        MaturityQuestionsComponent,
-        MaturityQuestionsAcetComponent,
-        MaturityQuestionsIseComponent,
-        AwwaStandardComponent,
-        DiagramQuestionsComponent,
-        SitesummaryCMMCComponent,
-        ExecutiveCMMCComponent,
-        ExecutiveCMMC2Component,
-        NavBackNextComponent,
-        CsetOriginComponent,
-        InherentRiskProfileComponent,
-        IrpSectionComponent,
-        AcetDonutChartComponent,
-        AcetExecutiveComponent,
-        AcetDeficencyComponent,
-        AcetCommentsmarkedComponent,
-        AcetAnsweredQuestionsComponent,
-        AcetCompensatingcontrolsComponent,
-        IseAnsweredQuestionsComponent,
-        IseMeritComponent,
-        IseExaminationComponent,
-        IseExaminerComponent,
-        IseDonutChartComponent,
-        IseDataComponent,
-        TutorialCmmcComponent,
-        TutorialEdmComponent,
-        TutorialRraComponent,
-        TutorialCrrComponent,
-        LoginAcetComponent,
-        LoginCsetComponent,
-        LoginRraComponent,
-        AboutCsetComponent,
-        AboutAcetComponent,
-        AcetOriginComponent,
-        GroupingBlockComponent,
-        QuestionBlockMaturityComponent,
-        QuestionBlockIseComponent,
-        EdmComponent,
-        EdmDeficiencyComponent,
-        GeneralDeficiencyComponent,
-        EdmCommentsmarkedComponent,
-        CisCommentsmarkedComponent,
-        QuestionTextComponent,
-        GlossaryTermComponent,
-        PlaceholderQuestionsComponent,
-        FeatureOptionComponent,
-        EdmRelationshipComponent,
-        EdmGlossaryComponent,
-        EdmIntroTextComponent,
-        EdmTocComponent,
-        EdmAcronymsComponent,
-        EdmSourceReferencesComponent,
-        EdmDomainDetailComponent,
-        RelationshipFormationComponent,
-        RelationshipManagementComponent,
-        ServiceProtectionComponent,
-        MaturityIndicatorLevelsComponent,
-        EDMTripleBarChart,
-        EDMHorizontalBarChart,
-        EDMBarChartLegend,
-        ModuleAddCloneComponent,
-        EdmPerfSummMil1Component,
-        EdmPerfSummAllMilComponent,
-        EDMFrameworkSummary,
-        EDMAppendixASectionOne,
-        EDMAppendixASectionTwo,
-        EdmBlocksCompactComponent,
-        EdmSummaryResultsComponent,
-        EdmBlocksCompactComponent,
-        EdmLegendSubquestionsComponent,
-        EdmPerfMil1Component,
-        EdmQBlocksHorizontalComponent,
-        EDMGoalQuestionSummary,
-        GroupingDescriptionComponent,
-        SummaryResultsComponent,
-        EDMGoalQuestionLegend,
-        CmmcDeficiencyComponent,
-        CmmcCommentsMarkedComponent,
-        CmmcAltJustificationsComponent,
-        CrrDeficiencyComponent,
-        CrrCommentsMarkedComponent,
-        RraGapsComponent,
-        RraLevelResultsComponent,
-        RraReportComponent,
-        RraDeficiencyComponent,
-        RraSummaryAllComponent,
-        RraSummaryComponent,
-        CommentsMfrComponent,
-        ReportDisclaimerComponent,
-        ReportAdvisoryComponent,
-        RraLevelsComponent,
-        RraAnswerCountsComponent,
-        RraAnswerDistributionComponent,
-        RraAnswerComplianceComponent,
-        RraQuestionsScoringComponent,
-        RraMiniUserGuideComponent,
-        IrpTabsComponent,
-        CrrSummaryResultsComponent,
-        CrrResultsPage,
-        CrrResultsDetailComponent,
-        CrrHeatmapComponent,
-        MatCommentsComponent,
-        AssessmentInfoTsaComponent,
-        TsaLayoutMainComponent,
-        TsaAssessmentCompleteComponent,
-        LoginTsaComponent,
-        AssessmentConfigTsaComponent,
-        FeatureOptionTsaComponent,
-        AboutTsaComponent,
-        TutorialCmmc2Component,
-        TopMenusComponent,
-        RraLayoutMainComponent,
-        AboutRraComponent,
-        LogoRraComponent,
-        IodLayoutComponent,
-        LogoCsetComponent,
-        TopMenusComponent,
-        LogoForReportsComponent,
-        LogoForReportsComponent,
-        QuestionBlockVadrComponent,
-        VadrDeficiencyComponent,
-        CsiComponent,
-        CsiOrganizationDemographicsComponent,
-        CsiServiceDemographicsComponent,
-        CsiServiceCompositionComponent,
-        AssessmentInfo2TsaComponent,
-        AssessmentDemographicsTsaComponent,
-        AssessmentComparisonAnalyticsComponent,
-        MaturityQuestionsNestedComponent,
-        QuestionBlockNestedComponent,
-        GroupingBlockNestedComponent,
-        OptionBlockNestedComponent,
-        ModuleContentComponent,
-        ModuleContentLaunchComponent,
-        TutorialCisComponent,
-        QuestionExtrasDialogComponent,
-        VadrReportComponent,
-        VadrAnswerComplianceComponent,
-        VadrAnswerCountsComponent,
-        VadrAnswerDistributionComponent,
-        VadrGapsComponent,
-        VadrLevelResultsComponent,
-        VadrLevelsComponent,
-        VadrQuestionsScoringComponent,
-        VadrSummaryComponent,
-        VadrSummaryAllComponent,
-        OpenEndedQuestionsComponent,
-        CisSurveyComponent,
-        GroupingBlockNestedReportComponent,
-        QuestionBlockNestedReportComponent,
-        OptionBlockNestedReportComponent,
-        AssessmentConfigNcuaComponent,
-        FeatureOptionNcuaComponent,
-        CoverSheetAComponent,
-        DisclaimerBlurbAComponent,
-        ConfigCisComponent,
-        CisRankedDeficiencyComponent,
-        RankedDeficienctyChartComponent,
-        CisCommentsmarkedComponent,
-        RankedDeficiencyComponent,
-        CisSectionScoringComponent,
-        CisScoringChartComponent,
-        SectionScoringComponent,
-        MergeExaminationsComponent,
-        CharterMismatchComponent,
-        DigitsOnlyNotZeroDirective,
-        LandingPageTabsComponent,
-        NewAssessmentComponent,
-        ModuleContentStandardComponent,
-        ModuleContentModelComponent,
-        McGroupingComponent,
-        McQuestionComponent,
-        McOptionComponent,
-        GuidanceBlockComponent,
-        ReferencesBlockComponent,
-        CrrReportComponent,
-        CrrCoverSheetComponent,
-        CrrCoverSheet2Component,
-        CrrIntroAboutComponent,
-        CrrMil1PerformanceSummaryComponent,
-        CrrPerformanceSummaryComponent,
-        CrrNistCsfSummaryComponent,
-        CrrMil1PerformanceComponent,
-        CrrResultsSummaryComponent,
-        CrrPercentageOfPracticesComponent,
-        CrrDomainDetailComponent,
-        CrrResourcesComponent,
-        CrrContactInformationComponent,
-        CrrAppendixACoverComponent,
-        CrrPerformanceAppendixAComponent,
-        CrrNistCsfCatSummaryComponent,
-        CrrNistCsfCatPerformanceComponent,
-        CmuNistCsfCatPerformanceComponent,
-        CrrSideTocComponent,
-        CmuPerformanceComponent,
-        CmuNistCsfCatSummaryComponent,
-        ImrReportComponent,
-        ReferencesBlockComponent,
-        NewAssessmentDialogComponent,
-        CrrMainTocComponent,
-        Cmmc2CommentsMarkedComponent,
-        Cmmc2DeficiencyComponent,
-        PrivacyWarningComponent,
-        PrivacyWarningRejectComponent,
-        SearchPageComponent,
-        LogoTsaComponent,
-        CfLayoutMainComponent,
-        RenewLayoutMainComponent,
-        OptionBlockComponent,
-        DemographicsExtendedComponent,
-        SectorHelpComponent,
-        AnalyticsCompareComponent,
-        AssessmentDetailCfComponent,
-        LoginCfComponent,
-        LoginRenewComponent,
-        MvraGapsComponent,
-        MvraSummaryComponent,
-        MvraAnswerFunctionsComponent,
-        MvraAnswerDomainsComponent,
-        MvraReportComponent,
-        AboutCfComponent,
-        MvraSummaryPageComponent,
-        MvraGapsPageComponent,
-        LoginAccessKeyComponent,
-        RelatedQBlockComponent,
-        CpgReportComponent,
-        CpgPracticeTableComponent,
-        RelatedQBlockComponent,
-        LogoutComponent,
-        CpgDomainSummaryComponent,
-        CpgCostImpactComplexityComponent,
-        CpgSummaryComponent,
-        CpgPracticesComponent,
-        OnlineDisclaimerComponent,
-        GlobalConfigurationComponent,
-        ReferencesTableComponent,
-        ReferencesDisplayComponent,
-        DiagramVulnerabilitiesDialogComponent,
-        DiagramVulnerabilitiesComponent,
-        AnalysisNetworkComponentsComponent,
-        SalSectionComponent,
-        AltJustificationCommentsComponent,
-        QuestionCommentsComponent,
-        QuestionsMarkedForReviewComponent,
-        StandardsComplianceComponent,
-        ComponentComplianceComponent,
-        OverallComplianceComponent,
-        RankedSubjectAreasComponent,
-        ComponentQuestionListComponent,
-        C2m2ReportComponent,
-        C2m2CoverSheetComponent,
-        C2m2IntroductionComponent,
-        C2m2ModelArchitectureComponent,
-        C2m2SummaryResultsComponent,
-        C2m2DetailedResultsComponent,
-        C2m2UsingSelfEvaluationResultsComponent,
-        C2m2SelfEvaluationNotesComponent,
-        C2m2ListOfPartiallyImplementedAndNotImplementedPracticesComponent,
-        C2m2SideTocComponent,
-        C2m2DonutComponent,
-        C2m2ObjectiveTableComponent,
-        CpgDomainSummaryTableComponent,
-        CpgDeficiencyComponent,
-        C2m2DomainMilBarChartComponent,
-        PdfReportsComponent,
-        InfoBlockComponent,
-        SiteInformationComponent,
-        AboutRenewComponent,
-        LogoCyberShieldComponent,
-        ExportPasswordComponent,
-        ImportPasswordComponent,
-        HydroDeficiencyComponent,
-        HydroDonutComponent,
-        HydroBarChartComponent,
-        HydroReportComponent,
-        HydroSideTocComponent,
-        HydroAcronymsComponent,
-        HydroExecutiveSummaryComponent,
-        HydroImportanceOfCybersecurityComponent,
-        HydroResultsSummaryComponent,
-        HydroConsequencesComponent,
-        HydroImpactCategoryComponent,
-        HydroFeasibilityReportComponent,
-        HydroBarVerticalComponent,
-        HydroImpactComponent,
-        HydroFeasibilityComponent,
-        HydroActionsComponent,
-        HydroActionItemComponent,
-        HydroProgressTotalsComponent,
-        HydroActionItemsReportComponent,
-        SdAnswerSummaryReportComponent,
-        SdAnswerSummaryComponent,
-        KeyReportComponent,
-        CmuGoalPerfStackedBarComponent,
-        CmuResultsDetailComponent,
-        CmuNistCsfSummaryComponent,
-        AssessmentConfigIodComponent,
-        AssessmentDemogIodComponent,
-        DemographicsIodComponent,
-        TutorialImrComponent,
-        ImrCoverSheetComponent,
-        ImrIntroAboutComponent,
-        ImrResourcesComponent,
-        CmuDomainComplianceComponent,
-        TsaSdComponent,
-        ImrSideTocComponent,
-        CmuAppendixCoverComponent,
-        OtherRemarksComponent,
-        CmuOtherRemarksComponent,
-        UserLanguageComponent,
-        MalcolmUploadErrorComponent,
-        AssessmentConvertCfComponent,
-        IseWarningsComponent,
-        TrendAnalyticsComponent
-    ],
-    providers: [
-        TranslocoService,
-        ConfigService,
-        AuthenticationService,
-        {
-            provide: APP_INITIALIZER,
-            useFactory: (configSvc: ConfigService, authSvc: AuthenticationService, tSvc: TranslocoService) => {
-                return () => {
-                    return configSvc.loadConfig().then(() => {
-                        // Load and set the language based on config
-                        return tSvc.load(configSvc.config.defaultLang).toPromise().then(() => {
-                            tSvc.setActiveLang(configSvc.config.defaultLang);
-                            return authSvc.checkLocal();
-                        });
-                    });
-                };
-            },
-            deps: [ConfigService, AuthenticationService, TranslocoService],
-            multi: true
-        },
-        {
-            provide: HTTP_INTERCEPTORS,
-            useClass: JwtInterceptor,
-            multi: true
-        },
-        DatePipe,
-        {
-            provide: MAT_DATE_LOCALE,
-            useFactory: (tSvc: TranslocoService) => {
-                // get the language based on config
-                return tSvc.getActiveLang();
-            },
-            deps: [TranslocoService],
-            multi: true
-        },
-        {
-            provide: DateAdapter,
-            useClass: MomentDateAdapter,
-            deps: [MAT_DATE_LOCALE, MAT_MOMENT_DATE_ADAPTER_OPTIONS]
-        },
-        {
-            provide: MAT_DATE_FORMATS, useValue: MAT_MOMENT_DATE_FORMATS
-        },
-        AuthGuard,
-        AssessGuard,
-        AggregationGuard,
-        DemographicService,
-        DemographicExtendedService,
-        AssessmentService,
-        EmailService,
-        QuestionsService,
-        SalService,
-        StandardService,
-        FrameworkService,
-        RequiredDocumentService,
-        IRPService,
-        ObservationsService,
-        NavigationService,
-        FileUploadClientService,
-        AnalysisService,
-        EnableFeatureService,
-        SetBuilderService,
-        ACETService,
-        ResourceLibraryService,
-        DiagramService,
-        AggregationService,
-        ChartService,
-        ChartColors,
-        ReportService,
-        ReportAnalysisService,
-        LocalStoreManager,
-        CmmcStyleService,
-        AcetFilteringService,
-        CmmcFilteringService,
-        EdmFilteringService,
-        CrrFilteringService,
-        RraFilteringService,
-        CrrService,
-        CmuService,
-        Utilities,
-        NCUAService,
-        GalleryService,
-        FooterService
-    ],
-    bootstrap: [AppComponent]
+  imports: [
+    BrowserModule,
+    BrowserAnimationsModule,
+    FormsModule,
+    HttpClientModule,
+    CommonModule,
+    AppRoutingModule,
+    // Material
+    A11yModule,
+    CdkAccordionModule,
+    ClipboardModule,
+    CdkStepperModule,
+    CdkTableModule,
+    CdkTreeModule,
+    DragDropModule,
+    MatAutocompleteModule,
+    MatBadgeModule,
+    MatBottomSheetModule,
+    MatButtonModule,
+    MatButtonToggleModule,
+    MatCardModule,
+    MatCheckboxModule,
+    MatChipsModule,
+    MatStepperModule,
+    MatDatepickerModule,
+    MatDialogModule,
+    MatDividerModule,
+    MatExpansionModule,
+    MatGridListModule,
+    MatIconModule,
+    MatInputModule,
+    MatListModule,
+    MatMenuModule,
+    MatNativeDateModule,
+    MatPaginatorModule,
+    MatProgressBarModule,
+    MatProgressSpinnerModule,
+    MatRadioModule,
+    MatRippleModule,
+    MatSelectModule,
+    MatSidenavModule,
+    MatSliderModule,
+    MatSlideToggleModule,
+    MatSnackBarModule,
+    MatSortModule,
+    MatTableModule,
+    MatTabsModule,
+    MatToolbarModule,
+    MatTooltipModule,
+    MatTreeModule,
+    OverlayModule,
+    PortalModule,
+    ScrollingModule,
+    // AutosizeModule,
+    // NgChartsModule,
+    // MatButtonModule,
+    // MatToolbarModule,
+    // MatChipsModule,
+    // MatSlideToggleModule,
+    // MatInputModule,
+    // MatCardModule,
+    // MatSliderModule,
+    // MatDatepickerModule,
+    // MatNativeDateModule,
+    // MatFormFieldModule,
+    // MatSortModule,
+    // MatExpansionModule,
+    // MatAutocompleteModule,
+    // MatDialogModule,
+    // MatTooltipModule,
+    // MatSnackBarModule,
+    // MatSidenavModule,
+    // MatTreeModule,
+    // MatIconModule,
+    // MatDividerModule,
+    // MatProgressSpinnerModule,
+    // MatProgressBarModule,
+    // MatListModule,
+    // MatMenuModule,
+    // MatTabsModule,
+    FlexLayoutModule,
+    ReactiveFormsModule,
+    NgxSliderModule,
+    // TextareaAutosizeModule,
+    FileUploadModule,
+    AngularEditorModule,
+    RouterModule,
+    CurrencyMaskModule,
+    TranslocoRootModule,
+    NgbModule,
+    NgxChartsModule,
+    TooltipModule,
+    SwiperModule,
+    EllipsisModule,
+    HotkeyModule.forRoot(),
+    CodeEditorModule.forRoot({
+      typingsWorkerUrl: 'assets/workers/typings-worker.js',
+      baseUrl: 'assets/monaco'
+    })
+  ],
+  declarations: [
+    AppComponent,
+    InitialComponent,
+    LoginComponent,
+    MyAssessmentsComponent,
+    AssessmentComponent,
+    ContactItemComponent,
+    PrepareComponent,
+    AssessmentInfoComponent,
+    AssessmentDetailComponent,
+    AssessmentContactsComponent,
+    AssessmentDemographicsComponent,
+    AssessmentInfoNcuaComponent,
+    AssessmentDetailNcuaComponent,
+    ResultsComponent,
+    SalSimpleComponent,
+    StandardsComponent,
+    ResetPassComponent,
+    EmailComponent,
+    ConfirmEqualValidatorDirective,
+    EmailValidatorDirective,
+    FocusDirective,
+    ZipCodeDirective,
+    AutoSizeDirective,
+    DigitsOnlyDirective,
+    RunScriptsDirective,
+    SalGenComponent,
+    SalNistComponent,
+    SalsComponent,
+    RegisterComponent,
+    EjectionComponent,
+    ChangePasswordComponent,
+    EditUserComponent,
+    AlertComponent,
+    ConfirmComponent,
+    FrameworkComponent,
+    RequiredDocsComponent,
+    IRPComponent,
+    ExamProfileComponent,
+    DiagramComponent,
+    AcetDetailComponent,
+    AboutComponent,
+    AdvisoryComponent,
+    QuestionsComponent,
+    QuestionBlockComponent,
+    QuestionExtrasComponent,
+    ResourceLibraryComponent,
+    OkayComponent,
+    ObservationsComponent,
+    IssuesComponent,
+    MeritCheckComponent,
+    SafePipe,
+    LinebreakPipe,
+    CompletionCountPipe,
+    LinebreakPlaintextPipe,
+    NullishCoalescePipe,
+    StatusCreateComponent,
+    ProgressComponent,
+    InViewComponent,
+    AnalysisComponent,
+    ReportsComponent,
+    OverviewComponent,
+    StandardsSummaryComponent,
+    StandardsRankedComponent,
+    StandardsResultsComponent,
+    ComponentsSummaryComponent,
+    ComponentsRankedComponent,
+    ComponentsResultsComponent,
+    ComponentsTypesComponent,
+    ComponentsWarningsComponent,
+    DashboardComponent,
+    RankedQuestionsComponent,
+    FeedbackComponent,
+    EnableProtectedComponent,
+    TermsOfUseComponent,
+    QuestionFiltersComponent,
+    AssessmentDocumentsComponent,
+    InlineParameterComponent,
+    GlobalParametersComponent,
+    ImportComponent,
+    UploadExportComponent,
+    KeyboardShortcutsComponent,
+    LicenseComponent,
+    AcetDashboardComponent,
+    AdminComponent,
+    SetListComponent,
+    CustomSetComponent,
+    RequirementListComponent,
+    QuestionListComponent,
+    BuilderBreadcrumbsComponent,
+    AddQuestionComponent,
+    RequirementDetailComponent,
+    AddRequirementComponent,
+    StandardDocumentsComponent,
+    RefDocumentComponent,
+    DomainMaturityFilterComponent,
+    IrpSummaryComponent,
+    ExamProfileSummaryComponent,
+    DiagramInventoryComponent,
+    DiagramInfoComponent,
+    DiagramComponentsComponent,
+    LinksComponent,
+    NetworkWarningsComponent,
+    ShapesComponent,
+    TextComponent,
+    ZonesComponent,
+    ExcelExportComponent,
+    NcuaExcelExportComponent,
+    MergeComponent,
+    MergeQuestionDetailComponent,
+    AggregationHomeComponent,
+    AliasAssessmentsComponent,
+    AggregationDetailComponent,
+    TrendAnalyticsComponent,
+    CompareAnalyticsComponent,
+    SelectAssessmentsComponent,
+    CompareSummaryComponent,
+    CompareMissedComponent,
+    CompareIndividualComponent,
+    CompareBestworstComponent,
+    ComponentOverrideComponent,
+    ExcelExportComponent,
+    LayoutBlankComponent,
+    LayoutSwitcherComponent,
+    LayoutMainComponent,
+    AcetLayoutMainComponent,
+    ReportTestComponent,
+    SiteDetailComponent,
+    ObservationTearoutsComponent,
+    EvalAgainstStandardsComponent,
+    ExecutiveSummaryComponent,
+    SecurityplanComponent,
+    SiteSummaryComponent,
+    TrendReportComponent,
+    TrendCompareCompatibilityComponent,
+    CompareReportComponent,
+    Assessment2InfoComponent,
+    ModelSelectComponent,
+    AssessmentConfigComponent,
+    CmmcAComponent,
+    CmmcLevelsComponent,
+    CmmcLevelResultsComponent,
+    CmmcLevelDrilldownComponent,
+    CmmcComplianceComponent,
+    CmmcGapsComponent,
+    Cmmc2LevelResultsComponent,
+    Cmmc2DomainResultsComponent,
+    SprsScoreComponent,
+    ComplianceScoreComponent,
+    ModelSelectComponent,
+    CategoryBlockComponent,
+    AskQuestionsComponent,
+    MaturityQuestionsComponent,
+    MaturityQuestionsAcetComponent,
+    MaturityQuestionsIseComponent,
+    AwwaStandardComponent,
+    DiagramQuestionsComponent,
+    SitesummaryCMMCComponent,
+    ExecutiveCMMCComponent,
+    ExecutiveCMMC2Component,
+    NavBackNextComponent,
+    CsetOriginComponent,
+    InherentRiskProfileComponent,
+    IrpSectionComponent,
+    AcetDonutChartComponent,
+    AcetExecutiveComponent,
+    AcetDeficencyComponent,
+    AcetCommentsmarkedComponent,
+    AcetAnsweredQuestionsComponent,
+    AcetCompensatingcontrolsComponent,
+    IseAnsweredQuestionsComponent,
+    IseMeritComponent,
+    IseExaminationComponent,
+    IseExaminerComponent,
+    IseDonutChartComponent,
+    IseDataComponent,
+    TutorialCmmcComponent,
+    TutorialEdmComponent,
+    TutorialRraComponent,
+    TutorialCrrComponent,
+    LoginAcetComponent,
+    LoginCsetComponent,
+    LoginRraComponent,
+    AboutCsetComponent,
+    AboutAcetComponent,
+    AcetOriginComponent,
+    GroupingBlockComponent,
+    QuestionBlockMaturityComponent,
+    QuestionBlockIseComponent,
+    EdmComponent,
+    EdmDeficiencyComponent,
+    GeneralDeficiencyComponent,
+    EdmCommentsmarkedComponent,
+    CisCommentsmarkedComponent,
+    QuestionTextComponent,
+    GlossaryTermComponent,
+    PlaceholderQuestionsComponent,
+    FeatureOptionComponent,
+    EdmRelationshipComponent,
+    EdmGlossaryComponent,
+    EdmIntroTextComponent,
+    EdmTocComponent,
+    EdmAcronymsComponent,
+    EdmSourceReferencesComponent,
+    EdmDomainDetailComponent,
+    RelationshipFormationComponent,
+    RelationshipManagementComponent,
+    ServiceProtectionComponent,
+    MaturityIndicatorLevelsComponent,
+    EDMTripleBarChart,
+    EDMHorizontalBarChart,
+    EDMBarChartLegend,
+    ModuleAddCloneComponent,
+    EdmPerfSummMil1Component,
+    EdmPerfSummAllMilComponent,
+    EDMFrameworkSummary,
+    EDMAppendixASectionOne,
+    EDMAppendixASectionTwo,
+    EdmBlocksCompactComponent,
+    EdmSummaryResultsComponent,
+    EdmBlocksCompactComponent,
+    EdmLegendSubquestionsComponent,
+    EdmPerfMil1Component,
+    EdmQBlocksHorizontalComponent,
+    EDMGoalQuestionSummary,
+    GroupingDescriptionComponent,
+    SummaryResultsComponent,
+    EDMGoalQuestionLegend,
+    CmmcDeficiencyComponent,
+    CmmcCommentsMarkedComponent,
+    CmmcAltJustificationsComponent,
+    CrrDeficiencyComponent,
+    CrrCommentsMarkedComponent,
+    RraGapsComponent,
+    RraLevelResultsComponent,
+    RraReportComponent,
+    RraDeficiencyComponent,
+    RraSummaryAllComponent,
+    RraSummaryComponent,
+    CommentsMfrComponent,
+    ReportDisclaimerComponent,
+    ReportAdvisoryComponent,
+    RraLevelsComponent,
+    RraAnswerCountsComponent,
+    RraAnswerDistributionComponent,
+    RraAnswerComplianceComponent,
+    RraQuestionsScoringComponent,
+    RraMiniUserGuideComponent,
+    IrpTabsComponent,
+    CrrSummaryResultsComponent,
+    CrrResultsPage,
+    CrrResultsDetailComponent,
+    CrrHeatmapComponent,
+    MatCommentsComponent,
+    AssessmentInfoTsaComponent,
+    TsaLayoutMainComponent,
+    TsaAssessmentCompleteComponent,
+    LoginTsaComponent,
+    AssessmentConfigTsaComponent,
+    FeatureOptionTsaComponent,
+    AboutTsaComponent,
+    TutorialCmmc2Component,
+    TopMenusComponent,
+    RraLayoutMainComponent,
+    AboutRraComponent,
+    LogoRraComponent,
+    IodLayoutComponent,
+    LogoCsetComponent,
+    TopMenusComponent,
+    LogoForReportsComponent,
+    LogoForReportsComponent,
+    QuestionBlockVadrComponent,
+    VadrDeficiencyComponent,
+    CsiComponent,
+    CsiOrganizationDemographicsComponent,
+    CsiServiceDemographicsComponent,
+    CsiServiceCompositionComponent,
+    AssessmentInfo2TsaComponent,
+    AssessmentDemographicsTsaComponent,
+    AssessmentComparisonAnalyticsComponent,
+    MaturityQuestionsNestedComponent,
+    QuestionBlockNestedComponent,
+    GroupingBlockNestedComponent,
+    OptionBlockNestedComponent,
+    ModuleContentComponent,
+    ModuleContentLaunchComponent,
+    TutorialCisComponent,
+    QuestionExtrasDialogComponent,
+    VadrReportComponent,
+    VadrAnswerComplianceComponent,
+    VadrAnswerCountsComponent,
+    VadrAnswerDistributionComponent,
+    VadrGapsComponent,
+    VadrLevelResultsComponent,
+    VadrLevelsComponent,
+    VadrQuestionsScoringComponent,
+    VadrSummaryComponent,
+    VadrSummaryAllComponent,
+    OpenEndedQuestionsComponent,
+    CisSurveyComponent,
+    GroupingBlockNestedReportComponent,
+    QuestionBlockNestedReportComponent,
+    OptionBlockNestedReportComponent,
+    AssessmentConfigNcuaComponent,
+    FeatureOptionNcuaComponent,
+    CoverSheetAComponent,
+    DisclaimerBlurbAComponent,
+    ConfigCisComponent,
+    CisRankedDeficiencyComponent,
+    RankedDeficienctyChartComponent,
+    CisCommentsmarkedComponent,
+    RankedDeficiencyComponent,
+    CisSectionScoringComponent,
+    CisScoringChartComponent,
+    SectionScoringComponent,
+    MergeExaminationsComponent,
+    CharterMismatchComponent,
+    DigitsOnlyNotZeroDirective,
+    LandingPageTabsComponent,
+    NewAssessmentComponent,
+    ModuleContentStandardComponent,
+    ModuleContentModelComponent,
+    McGroupingComponent,
+    McQuestionComponent,
+    McOptionComponent,
+    GuidanceBlockComponent,
+    ReferencesBlockComponent,
+    CrrReportComponent,
+    CrrCoverSheetComponent,
+    CrrCoverSheet2Component,
+    CrrIntroAboutComponent,
+    CrrMil1PerformanceSummaryComponent,
+    CrrPerformanceSummaryComponent,
+    CrrNistCsfSummaryComponent,
+    CrrMil1PerformanceComponent,
+    CrrResultsSummaryComponent,
+    CrrPercentageOfPracticesComponent,
+    CrrDomainDetailComponent,
+    CrrResourcesComponent,
+    CrrContactInformationComponent,
+    CrrAppendixACoverComponent,
+    CrrPerformanceAppendixAComponent,
+    CrrNistCsfCatSummaryComponent,
+    CrrNistCsfCatPerformanceComponent,
+    CmuNistCsfCatPerformanceComponent,
+    CrrSideTocComponent,
+    CmuPerformanceComponent,
+    CmuNistCsfCatSummaryComponent,
+    ImrReportComponent,
+    ReferencesBlockComponent,
+    NewAssessmentDialogComponent,
+    CrrMainTocComponent,
+    Cmmc2CommentsMarkedComponent,
+    Cmmc2DeficiencyComponent,
+    PrivacyWarningComponent,
+    PrivacyWarningRejectComponent,
+    SearchPageComponent,
+    LogoTsaComponent,
+    CfLayoutMainComponent,
+    RenewLayoutMainComponent,
+    OptionBlockComponent,
+    DemographicsExtendedComponent,
+    SectorHelpComponent,
+    AnalyticsCompareComponent,
+    AssessmentDetailCfComponent,
+    LoginCfComponent,
+    LoginRenewComponent,
+    MvraGapsComponent,
+    MvraSummaryComponent,
+    MvraAnswerFunctionsComponent,
+    MvraAnswerDomainsComponent,
+    MvraReportComponent,
+    AboutCfComponent,
+    MvraSummaryPageComponent,
+    MvraGapsPageComponent,
+    LoginAccessKeyComponent,
+    RelatedQBlockComponent,
+    CpgReportComponent,
+    CpgPracticeTableComponent,
+    RelatedQBlockComponent,
+    LogoutComponent,
+    CpgDomainSummaryComponent,
+    CpgCostImpactComplexityComponent,
+    CpgSummaryComponent,
+    CpgPracticesComponent,
+    OnlineDisclaimerComponent,
+    GlobalConfigurationComponent,
+    ReferencesTableComponent,
+    ReferencesDisplayComponent,
+    DiagramVulnerabilitiesDialogComponent,
+    DiagramVulnerabilitiesComponent,
+    AnalysisNetworkComponentsComponent,
+    SalSectionComponent,
+    AltJustificationCommentsComponent,
+    QuestionCommentsComponent,
+    QuestionsMarkedForReviewComponent,
+    StandardsComplianceComponent,
+    ComponentComplianceComponent,
+    OverallComplianceComponent,
+    RankedSubjectAreasComponent,
+    ComponentQuestionListComponent,
+    C2m2ReportComponent,
+    C2m2CoverSheetComponent,
+    C2m2IntroductionComponent,
+    C2m2ModelArchitectureComponent,
+    C2m2SummaryResultsComponent,
+    C2m2DetailedResultsComponent,
+    C2m2UsingSelfEvaluationResultsComponent,
+    C2m2SelfEvaluationNotesComponent,
+    C2m2ListOfPartiallyImplementedAndNotImplementedPracticesComponent,
+    C2m2SideTocComponent,
+    C2m2DonutComponent,
+    C2m2ObjectiveTableComponent,
+    CpgDomainSummaryTableComponent,
+    CpgDeficiencyComponent,
+    C2m2DomainMilBarChartComponent,
+    PdfReportsComponent,
+    InfoBlockComponent,
+    SiteInformationComponent,
+    AboutRenewComponent,
+    LogoCyberShieldComponent,
+    ExportPasswordComponent,
+    ImportPasswordComponent,
+    HydroDeficiencyComponent,
+    HydroDonutComponent,
+    HydroBarChartComponent,
+    HydroReportComponent,
+    HydroSideTocComponent,
+    HydroAcronymsComponent,
+    HydroExecutiveSummaryComponent,
+    HydroImportanceOfCybersecurityComponent,
+    HydroResultsSummaryComponent,
+    HydroConsequencesComponent,
+    HydroImpactCategoryComponent,
+    HydroFeasibilityReportComponent,
+    HydroBarVerticalComponent,
+    HydroImpactComponent,
+    HydroFeasibilityComponent,
+    HydroActionsComponent,
+    HydroActionItemComponent,
+    HydroProgressTotalsComponent,
+    HydroActionItemsReportComponent,
+    SdAnswerSummaryReportComponent,
+    SdAnswerSummaryComponent,
+    KeyReportComponent,
+    CmuGoalPerfStackedBarComponent,
+    CmuResultsDetailComponent,
+    CmuNistCsfSummaryComponent,
+    AssessmentConfigIodComponent,
+    AssessmentDemogIodComponent,
+    DemographicsIodComponent,
+    TutorialImrComponent,
+    ImrCoverSheetComponent,
+    ImrIntroAboutComponent,
+    ImrResourcesComponent,
+    CmuDomainComplianceComponent,
+    TsaSdComponent,
+    ImrSideTocComponent,
+    CmuAppendixCoverComponent,
+    OtherRemarksComponent,
+    CmuOtherRemarksComponent,
+    UserLanguageComponent,
+    MalcolmUploadErrorComponent,
+    AssessmentConvertCfComponent,
+    IseWarningsComponent,
+    TrendAnalyticsComponent
+  ],
+  providers: [
+    TranslocoService,
+    ConfigService,
+    AuthenticationService,
+    {
+      provide: APP_INITIALIZER,
+      useFactory: (configSvc: ConfigService, authSvc: AuthenticationService, tSvc: TranslocoService) => {
+        return () => {
+          return configSvc.loadConfig().then(() => {
+            // Load and set the language based on config
+            return tSvc
+              .load(configSvc.config.defaultLang)
+              .toPromise()
+              .then(() => {
+                tSvc.setActiveLang(configSvc.config.defaultLang);
+                return authSvc.checkLocal();
+              });
+          });
+        };
+      },
+      deps: [ConfigService, AuthenticationService, TranslocoService],
+      multi: true
+    },
+    {
+      provide: HTTP_INTERCEPTORS,
+      useClass: JwtInterceptor,
+      multi: true
+    },
+    DatePipe,
+    {
+      provide: MAT_DATE_LOCALE,
+      useFactory: (tSvc: TranslocoService) => {
+        // get the language based on config
+        return tSvc.getActiveLang();
+      },
+      deps: [TranslocoService],
+      multi: true
+    },
+    {
+      provide: DateAdapter,
+      useClass: MomentDateAdapter,
+      deps: [MAT_DATE_LOCALE, MAT_MOMENT_DATE_ADAPTER_OPTIONS]
+    },
+    {
+      provide: MAT_DATE_FORMATS,
+      useValue: MAT_MOMENT_DATE_FORMATS
+    },
+    AuthGuard,
+    AssessGuard,
+    AggregationGuard,
+    DemographicService,
+    DemographicExtendedService,
+    AssessmentService,
+    EmailService,
+    QuestionsService,
+    SalService,
+    StandardService,
+    FrameworkService,
+    RequiredDocumentService,
+    IRPService,
+    ObservationsService,
+    NavigationService,
+    FileUploadClientService,
+    AnalysisService,
+    EnableFeatureService,
+    SetBuilderService,
+    ACETService,
+    ResourceLibraryService,
+    DiagramService,
+    AggregationService,
+    ChartService,
+    ChartColors,
+    ReportService,
+    ReportAnalysisService,
+    LocalStoreManager,
+    CmmcStyleService,
+    AcetFilteringService,
+    CmmcFilteringService,
+    EdmFilteringService,
+    CrrFilteringService,
+    RraFilteringService,
+    CmuService,
+    CmuService,
+    Utilities,
+    NCUAService,
+    GalleryService,
+    FooterService
+  ],
+  bootstrap: [AppComponent]
 })
-
-export class AppModule { }
+export class AppModule {}
