+++ CSETWebNg/src/app/assessment/results/crr/crr-summary-results/crr-summary-results.component.ts
@@ -22,28 +22,24 @@
 //
 ////////////////////////////////
 import { Component, OnInit } from '@angular/core';
-import { CrrService } from '../../../../services/crr.service';
-import { CrrReportModel } from '../../../../models/reports.model';
+import { CmuService } from '../../../../services/cmu.service';
+import { CmuReportModel } from '../../../../models/reports.model';
 
 @Component({
   selector: 'app-crr-summary-results',
   templateUrl: './crr-summary-results.component.html'
 })
 export class CrrSummaryResultsComponent implements OnInit {
-
   summaryResult: any = '';
-  crrModel: CrrReportModel;
+  crrModel: CmuReportModel;
 
   chartLoaded: boolean = false;
   summaryResultLoaded: boolean = false;
 
-  constructor(
-    private crrSvc: CrrService,
-  ) {
-  }
+  constructor(private cmuSvc: CmuService) {}
 
   ngOnInit(): void {
-    this.crrSvc.getCrrModel().subscribe((data: CrrReportModel) => {
+    this.cmuSvc.getCmuModel().subscribe((data: CmuReportModel) => {
       this.crrModel = data;
       this.chartLoaded = true;
       this.summaryResultLoaded = true;
