+++ CSETWebNg/src/app/reports/crr/crr-comments-marked/crr-comments-marked.component.ts
@@ -21,8 +21,8 @@
 //  SOFTWARE.
 //
 ////////////////////////////////
-import { CrrReportModel } from './../../../models/reports.model';
-import { CrrService } from './../../../services/crr.service';
+import { CmuReportModel } from './../../../models/reports.model';
+import { CmuService } from './../../../services/cmu.service';
 import { Component, OnInit } from '@angular/core';
 import { Title } from '@angular/platform-browser';
 import { ConfigService } from '../../../services/config.service';
@@ -34,8 +34,7 @@ import { QuestionsService } from '../../../services/questions.service';
   styleUrls: ['./../crr-report/crr-report.component.scss']
 })
 export class CrrCommentsMarkedComponent implements OnInit {
-
-  crrModel: CrrReportModel;
+  crrModel: CmuReportModel;
   loading: boolean = false;
   keyToCategory: any;
 
@@ -45,26 +44,26 @@ export class CrrCommentsMarkedComponent implements OnInit {
   constructor(
     public configSvc: ConfigService,
     private titleService: Title,
-    private crrSvc: CrrService,
+    private cmuSvc: CmuService,
     public questionsSvc: QuestionsService
-  ) { }
+  ) {}
 
   ngOnInit(): void {
     this.loading = true;
-    this.titleService.setTitle("Comments Report - CISA CRR");
-    this.keyToCategory = this.crrSvc.keyToCategory;
+    this.titleService.setTitle('Comments Report - CISA CRR');
+    this.keyToCategory = this.cmuSvc.keyToCategory;
 
-    let appCode = this.configSvc.installationMode;
+    const appCode = this.configSvc.installationMode;
 
-    this.crrSvc.getCrrModel().subscribe(
-      (r: CrrReportModel) => {
+    this.cmuSvc.getCmuModel().subscribe(
+      (r: CmuReportModel) => {
         this.crrModel = r;
-        let commentsCategories = [];
+        const commentsCategories = [];
 
         // Build up comments list
-        this.crrModel.reportData.comments.forEach(matAns => {
+        this.crrModel.reportData.comments.forEach((matAns) => {
           const domain = matAns.mat.question_Title.split(':')[0];
-          const cElement = commentsCategories.find(e => e.cat === this.keyToCategory[domain]);
+          const cElement = commentsCategories.find((e) => e.cat === this.keyToCategory[domain]);
 
           if (!cElement) {
             commentsCategories.push({ cat: this.keyToCategory[domain], matAnswers: [matAns] });
@@ -74,26 +73,29 @@ export class CrrCommentsMarkedComponent implements OnInit {
         });
 
         // We want the report to have the categories in a particular order, so manually add them in order
-        this.pushCommentsCategory(commentsCategories.find(e => e.cat === this.keyToCategory['AM']));
-        this.pushCommentsCategory(commentsCategories.find(e => e.cat === this.keyToCategory['CM']));
-        this.pushCommentsCategory(commentsCategories.find(e => e.cat === this.keyToCategory['CCM']));
-        this.pushCommentsCategory(commentsCategories.find(e => e.cat === this.keyToCategory['VM']));
-        this.pushCommentsCategory(commentsCategories.find(e => e.cat === this.keyToCategory['IM']));
-        this.pushCommentsCategory(commentsCategories.find(e => e.cat === this.keyToCategory['SCM']));
-        this.pushCommentsCategory(commentsCategories.find(e => e.cat === this.keyToCategory['RM']));
-        this.pushCommentsCategory(commentsCategories.find(e => e.cat === this.keyToCategory['EDM']));
-        this.pushCommentsCategory(commentsCategories.find(e => e.cat === this.keyToCategory['TA']));
-        this.pushCommentsCategory(commentsCategories.find(e => e.cat === this.keyToCategory['SA']));
+        this.pushCommentsCategory(commentsCategories.find((e) => e.cat === this.keyToCategory['AM']));
+        this.pushCommentsCategory(commentsCategories.find((e) => e.cat === this.keyToCategory['CM']));
+        this.pushCommentsCategory(commentsCategories.find((e) => e.cat === this.keyToCategory['CCM']));
+        this.pushCommentsCategory(commentsCategories.find((e) => e.cat === this.keyToCategory['VM']));
+        this.pushCommentsCategory(commentsCategories.find((e) => e.cat === this.keyToCategory['IM']));
+        this.pushCommentsCategory(commentsCategories.find((e) => e.cat === this.keyToCategory['SCM']));
+        this.pushCommentsCategory(commentsCategories.find((e) => e.cat === this.keyToCategory['RM']));
+        this.pushCommentsCategory(commentsCategories.find((e) => e.cat === this.keyToCategory['EDM']));
+        this.pushCommentsCategory(commentsCategories.find((e) => e.cat === this.keyToCategory['TA']));
+        this.pushCommentsCategory(commentsCategories.find((e) => e.cat === this.keyToCategory['SA']));
 
         // Sort the marked for review list
-        this.commentsList.forEach(e => {
+        this.commentsList.forEach((e) => {
           e.matAnswers.sort((a, b) => {
-            return a.mat.question_Title.split('-')[0].localeCompare(b.mat.question_Title.split('-')[0]) || a.mat.question_Text.localeCompare(b.mat.question_Text);
-          })
+            return (
+              a.mat.question_Title.split('-')[0].localeCompare(b.mat.question_Title.split('-')[0]) ||
+              a.mat.question_Text.localeCompare(b.mat.question_Text)
+            );
+          });
         });
 
         // mark questions followed by a child for border display
-        this.commentsList.forEach(e => {
+        this.commentsList.forEach((e) => {
           for (let i = 0; i < e.matAnswers.length; i++) {
             if (e.matAnswers[i + 1]?.mat.parent_Question_Id != null) {
               e.matAnswers[i].isFollowedByChild = true;
@@ -101,12 +103,12 @@ export class CrrCommentsMarkedComponent implements OnInit {
           }
         });
 
-        let mfrCategories = [];
+        const mfrCategories = [];
 
         // Build up marked for review list
-        this.crrModel.reportData.markedForReviewList.forEach(matAns => {
+        this.crrModel.reportData.markedForReviewList.forEach((matAns) => {
           const domain = matAns.mat.question_Title.split(':')[0];
-          const mfrElement = mfrCategories.find(e => e.cat === this.keyToCategory[domain]);
+          const mfrElement = mfrCategories.find((e) => e.cat === this.keyToCategory[domain]);
 
           if (!mfrElement) {
             mfrCategories.push({ cat: this.keyToCategory[domain], matAnswers: [matAns] });
@@ -116,26 +118,29 @@ export class CrrCommentsMarkedComponent implements OnInit {
         });
 
         // We want the report to have the categories in a particular order, so manually add them in order
-        this.pushMfrCategory(mfrCategories.find(e => e.cat === this.keyToCategory['AM']));
-        this.pushMfrCategory(mfrCategories.find(e => e.cat === this.keyToCategory['CM']));
-        this.pushMfrCategory(mfrCategories.find(e => e.cat === this.keyToCategory['CCM']));
-        this.pushMfrCategory(mfrCategories.find(e => e.cat === this.keyToCategory['VM']));
-        this.pushMfrCategory(mfrCategories.find(e => e.cat === this.keyToCategory['IM']));
-        this.pushMfrCategory(mfrCategories.find(e => e.cat === this.keyToCategory['SCM']));
-        this.pushMfrCategory(mfrCategories.find(e => e.cat === this.keyToCategory['RM']));
-        this.pushMfrCategory(mfrCategories.find(e => e.cat === this.keyToCategory['EDM']));
-        this.pushMfrCategory(mfrCategories.find(e => e.cat === this.keyToCategory['TA']));
-        this.pushMfrCategory(mfrCategories.find(e => e.cat === this.keyToCategory['SA']));
+        this.pushMfrCategory(mfrCategories.find((e) => e.cat === this.keyToCategory['AM']));
+        this.pushMfrCategory(mfrCategories.find((e) => e.cat === this.keyToCategory['CM']));
+        this.pushMfrCategory(mfrCategories.find((e) => e.cat === this.keyToCategory['CCM']));
+        this.pushMfrCategory(mfrCategories.find((e) => e.cat === this.keyToCategory['VM']));
+        this.pushMfrCategory(mfrCategories.find((e) => e.cat === this.keyToCategory['IM']));
+        this.pushMfrCategory(mfrCategories.find((e) => e.cat === this.keyToCategory['SCM']));
+        this.pushMfrCategory(mfrCategories.find((e) => e.cat === this.keyToCategory['RM']));
+        this.pushMfrCategory(mfrCategories.find((e) => e.cat === this.keyToCategory['EDM']));
+        this.pushMfrCategory(mfrCategories.find((e) => e.cat === this.keyToCategory['TA']));
+        this.pushMfrCategory(mfrCategories.find((e) => e.cat === this.keyToCategory['SA']));
 
         // Sort the marked for review list
-        this.markedForReviewList.forEach(e => {
+        this.markedForReviewList.forEach((e) => {
           e.matAnswers.sort((a, b) => {
-            return a.mat.question_Title.split('-')[0].localeCompare(b.mat.question_Title.split('-')[0]) || a.mat.question_Text.localeCompare(b.mat.question_Text);
-          })
+            return (
+              a.mat.question_Title.split('-')[0].localeCompare(b.mat.question_Title.split('-')[0]) ||
+              a.mat.question_Text.localeCompare(b.mat.question_Text)
+            );
+          });
         });
 
         // mark questions followed by a child for border display
-        this.markedForReviewList.forEach(e => {
+        this.markedForReviewList.forEach((e) => {
           for (let i = 0; i < e.matAnswers.length; i++) {
             if (e.matAnswers[i + 1]?.mat.parent_Question_Id != null) {
               e.matAnswers[i].isFollowedByChild = true;
@@ -145,7 +150,7 @@ export class CrrCommentsMarkedComponent implements OnInit {
 
         this.loading = false;
       },
-      error => console.log('CRR Comments Marked Report Error: ' + (<Error>error).message)
+      (error) => console.log('CRR Comments Marked Report Error: ' + (<Error>error).message)
     );
   }
