+++ CSETWebNg/src/app/reports/crr/crr-report/crr-cover-sheet2/crr-cover-sheet2.component.ts
@@ -23,7 +23,7 @@
 ////////////////////////////////
 import { Component, Input, OnInit } from '@angular/core';
 import { ConfigService } from '../../../../services/config.service';
-import { CrrReportModel } from '../../../../models/reports.model';
+import { CmuReportModel } from '../../../../models/reports.model';
 
 @Component({
   selector: 'app-crr-cover-sheet2',
@@ -31,19 +31,17 @@ import { CrrReportModel } from '../../../../models/reports.model';
   styleUrls: ['./../crr-report.component.scss']
 })
 export class CrrCoverSheet2Component implements OnInit {
-
-  @Input() model: CrrReportModel;
+  @Input() model: CmuReportModel;
 
   logoUrl: string;
 
-  constructor(private configSvc: ConfigService) { }
+  constructor(private configSvc: ConfigService) {}
 
   ngOnInit(): void {
     if (this.configSvc.installationMode === 'TSA') {
       this.logoUrl = 'assets/images/TSA/tsa_insignia_rgbtransparent.png';
     } else {
-      this.logoUrl = 'assets/images/CISA_Logo_1831px.png'
+      this.logoUrl = 'assets/images/CISA_Logo_1831px.png';
     }
   }
-
 }
