+++ CSETWebNg/src/app/reports/crr/crr-report/crr-nist-csf-cat-performance/crr-nist-csf-cat-performance.component.ts
@@ -22,8 +22,8 @@
 //
 ////////////////////////////////
 import { Component, Input, OnInit } from '@angular/core';
-import { CrrReportModel } from '../../../../models/reports.model';
-import { CrrService } from '../../../../services/crr.service';
+import { CmuReportModel } from '../../../../models/reports.model';
+import { CmuService } from '../../../../services/cmu.service';
 
 @Component({
   selector: 'app-crr-nist-csf-cat-performance',
@@ -31,17 +31,15 @@ import { CrrService } from '../../../../services/crr.service';
   styleUrls: ['./../crr-report.component.scss']
 })
 export class CrrNistCsfCatPerformanceComponent implements OnInit {
-
-  @Input() model: CrrReportModel;
+  @Input() model: CmuReportModel;
 
   bodyData: any[] = [];
 
-  constructor(private crrSvc: CrrService) { }
+  constructor(private cmuSvc: CmuService) {}
 
   ngOnInit(): void {
-    this.crrSvc.getNistCsfCatPerformanceBodyData().subscribe((resp: any[]) => {
+    this.cmuSvc.getNistCsfCatPerformanceBodyData().subscribe((resp: any[]) => {
       this.bodyData = resp;
-    })
+    });
   }
-
 }
