+++ CSETWebNg/src/app/reports/crr/crr-report/crr-report.component.html
@@ -47,51 +47,51 @@
             <div class="report-body-crr">
 
               <div id="toc-top">
-                <app-crr-cover-sheet [model]="crrModel"></app-crr-cover-sheet>
+                <app-crr-cover-sheet [model]="cmuModel"></app-crr-cover-sheet>
               </div>
 
               <div class="page-margin-top">
-                <app-crr-cover-sheet2 [model]="crrModel"></app-crr-cover-sheet2>
+                <app-crr-cover-sheet2 [model]="cmuModel"></app-crr-cover-sheet2>
               </div>
 
               <div id="main-toc" class="page-margin-top">
-                <app-crr-main-toc [model]="crrModel"></app-crr-main-toc>
+                <app-crr-main-toc [model]="cmuModel"></app-crr-main-toc>
               </div>
 
               <div id="toc-intro" class="page-margin-top">
-                <app-crr-intro-about [model]="crrModel"></app-crr-intro-about>
+                <app-crr-intro-about [model]="cmuModel"></app-crr-intro-about>
               </div>
 
               <div id="toc-mil1-ps" class="page-margin-top">
-                <app-crr-mil1-performance-summary [model]="crrModel"></app-crr-mil1-performance-summary>
+                <app-crr-mil1-performance-summary [model]="cmuModel"></app-crr-mil1-performance-summary>
               </div>
 
               <div id="toc-crr-ps" class="page-margin-top">
-                <app-crr-performance-summary [model]="crrModel"></app-crr-performance-summary>
+                <app-crr-performance-summary [model]="cmuModel"></app-crr-performance-summary>
               </div>
 
               <div id="toc-nist-csf" class="page-margin-top">
-                <app-crr-nist-csf-summary [model]="crrModel"></app-crr-nist-csf-summary>
+                <app-crr-nist-csf-summary [model]="cmuModel"></app-crr-nist-csf-summary>
               </div>
 
               <div id="toc-mil1-performance" class="page-margin-top">
-                <app-crr-mil1-performance [model]="crrModel"></app-crr-mil1-performance>
+                <app-crr-mil1-performance [model]="cmuModel"></app-crr-mil1-performance>
               </div>
 
               <div id="toc-summary-results" class="page-margin-top" style="padding-bottom: 60px">
-                <app-crr-results-summary [model]="crrModel"></app-crr-results-summary>
+                <app-crr-results-summary [model]="cmuModel"></app-crr-results-summary>
               </div>
 
               <div id="toc-percentage-practices" class="page-margin-top">
-                <app-crr-percentage-of-practices [model]="crrModel"></app-crr-percentage-of-practices>
+                <app-crr-percentage-of-practices [model]="cmuModel"></app-crr-percentage-of-practices>
               </div>
 
               <div id="toc-domain-detail" class="page-margin-top">
-                <app-crr-domain-detail [model]="crrModel"></app-crr-domain-detail>
+                <app-crr-domain-detail [model]="cmuModel"></app-crr-domain-detail>
               </div>
 
               <div id="toc-resources-referenced" class="page-margin-top">
-                <app-crr-resources [model]="crrModel"></app-crr-resources>
+                <app-crr-resources [model]="cmuModel"></app-crr-resources>
               </div>
 
               <div id="toc-contact-information" class="page-margin-top">
@@ -103,15 +103,15 @@
               </div>
 
               <div id="toc-crr-performance-app-a" class="page-margin-top">
-                <app-crr-performance-appendix-a [model]="crrModel"></app-crr-performance-appendix-a>
+                <app-crr-performance-appendix-a [model]="cmuModel"></app-crr-performance-appendix-a>
               </div>
 
               <div id="toc-nist-framework-summary-app-a" class="page-margin-top">
-                <app-crr-nist-csf-cat-summary [model]="crrModel"></app-crr-nist-csf-cat-summary>
+                <app-crr-nist-csf-cat-summary [model]="cmuModel"></app-crr-nist-csf-cat-summary>
               </div>
 
               <div id="toc-nist-framework-performance-app-a" class="page-margin-top">
-                <app-crr-nist-csf-cat-performance [model]="crrModel"></app-crr-nist-csf-cat-performance>
+                <app-crr-nist-csf-cat-performance [model]="cmuModel"></app-crr-nist-csf-cat-performance>
               </div>
 
             </div>
