+++ CSETWebNg/src/app/reports/crr/crr-report/crr-results-summary/crr-results-summary.component.html
@@ -35,7 +35,7 @@
         </div>
         <div class="CrrResultsBody">
             <div class="crrResultsTable">
-                <div *ngFor="let domain of model?.crrResultsData.crrDomains" class="crrResultsTableRow">
+                <div *ngFor="let domain of model?.cmuResultsData.crrDomains" class="crrResultsTableRow">
                     <div class="crrResultsDomainCell">
                         <div class="crrResultsTableDomainName">
                             {{ domain.domainName }}
